+++ app/frontend/packs/application.js
@@ -23,6 +23,9 @@ import '../scripts/shared/post'
 import '../scripts/shared/tooltip'
 import '../scripts/shared/mytask'
 import '../scripts/shared/githubLink'
+import '../scripts/shared/workspaceArrow'
+import '../scripts/shared/chatroomIsactive'
+import '../scripts/shared/itemStatus'
 import 'scripts'
 import 'stylesheets'
 import '../stylesheets/index'
