+++ app/views/webrtc/index.html.erb
@@ -1,5 +1,6 @@
 <div class="container-all">
   <div class="webcam-btn">
+    <i class="fas fa-chevron-circle-left workspace-list-colse"></i>
     <button id="join-button">Join</button>
     <button id="leave-button">Leave</button>
     <button id="Start-button">Start Cam</button>
@@ -8,13 +9,13 @@
     <button id="not-share-button">Stop Desktop</button>
     <button id="audio-button">audio</button>
   </div>
-  <div class="meeting-room">
+  <div class="meeting-room" id="meeting-room">
     <div class="container-for-webcam">
       <div class="attendees" id="remote-video-container">
       </div>
       <video class="local-video" id="local-video" autoplay></video>
     </div>
-    <div class="text-room">123</div>
+    <%# <div class="text-room">123</div> %>
   </div>
 </div>
